+++ ceammc/ext/src/array/array_fill.h
@@ -27,6 +27,7 @@ public:
     void onList(const AtomList& l);
 
     void m_fill(t_symbol* m, const AtomList& l);
+    void m_sin(t_symbol* m, const AtomList& l);
     AtomList parseRange(const AtomList& args, size_t* from, size_t* to) const;
 
 private: